requirements.txt
asttokens==3.0.0
cloudpickle==3.1.1
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.12
decorator==5.1.1
executing==2.2.0
faiss-cpu==1.10.0
fonttools==4.56.0
imageio==2.37.0
ipykernel==6.29.5
ipython==8.32.0
jedi==0.19.2
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
lazy_loader==0.4
lime==0.2.0.1
llvmlite==0.44.0
matplotlib==3.10.0
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
networkx==3.4.2
numba==0.61.0
numpy==2.1.3
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
prompt_toolkit==3.0.50
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2025.1
pyzmq==26.2.1
scikit-image==0.25.1
scikit-learn==1.6.1
scipy==1.15.2
shap==0.46.0
six==1.17.0
slicer==0.0.8
stack-data==0.6.3
threadpoolctl==3.5.0
tifffile==2025.1.10
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
tzdata==2025.1
wcwidth==0.2.13
